backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "groq>=1.0.0",
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "mistralai>=1.10.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pymupdf>=1.26.7",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "reportlab>=4.4.9",
    "sentence-transformers>=5.2.0",
    "supabase>=2.27.2",
    "uvicorn>=0.40.0",
]
